Learn to prepare data for your next machine learning project
Text cleaning and processing is an important task in every machine learning project where the task is to make sense of textual data. How to construct features from Text Data and further to it, create synthetic features are again critical tasks. On top of it how to apply machine learning models to create classifiers are also difficult.