Ecommerce product reviews - Pairwise ranking and sentiment analysis
This project analyzes a dataset containing ecommerce product reviews. The goal is to use machine learning models to perform sentiment analysis on product reviews and rank them based on relevance. Reviews play a key role in product recommendation systems.


E-Commerce applications provide an added advantage to customers to buy a product with added suggestions in the form of reviews. Obviously, reviews are useful and impactful for customers who are going to buy the products. But these enormous amounts of reviews also create problems for customers as they are not able to segregate useful ones. Regardless, these immense proportions of reviews make an issue for customers as it becomes very difficult to filter informative reviews. This proportional issue has been attempted in this project. The approach that we discuss in detail later ranks reviews based on their relevance with the product and rank down irrelevant reviews.

This work has been done in four phases- data preprocessing/filtering (which includes Language Detection, Gibberish Detection, Profanity Detection), feature extraction, pairwise review ranking, and classification. The outcome will be a list of reviews for a particular product ranking on the basis of relevance using a pairwise ranking approach.